Sesion, tabs error

+++ app/controllers/ControllerEmpresa.java
@@ -30,28 +30,34 @@ public class ControllerEmpresa extends Controller{
      PdfGenerator pdfGenerator;
      
       public Result document() {
-         List<FormularioEmpresa> formEmp = FormularioEmpresa.find.all();
-         RegistroEmpresa registro = new RegistroEmpresa();
-        return pdfGenerator.ok(perfilEmpresa.render("Pdf empresa",formEmp,registro), "http://localhost:9000").as("application/pdf");
-    }
+            List<FormularioEmpresa> formEmp = FormularioEmpresa.find.all();
+            RegistroEmpresa registro = new RegistroEmpresa();
+           return pdfGenerator.ok(perfilEmpresa.render("Pdf empresa",formEmp,registro), "http://localhost:9000").as("application/pdf");//content type
+        }
      //METODO PARA LISTAR FORMULARIO EMPRESAS
     public Result listarFormularioEmpresa(){
         RegistroEmpresa empresa = new EmpresaSession().getRegistroEmpresa();
        // List<FormularioEmpresa> formulario = FormularioEmpresa.find.all();
-        List<FormularioEmpresa> formEmpresa = FormularioEmpresa.find.where().ilike("registroEmpresa", ""+empresa.id).findList();
-        System.out.println("");
-        System.out.println("id: "+empresa.id);
-        System.err.println("TAM "+formEmpresa.size());
-        return ok(perfilEmpresa.render("", formEmpresa, empresa));
+       if(empresa!=null){
+            List<FormularioEmpresa> formEmpresa = FormularioEmpresa.find.where().ilike("registroEmpresa", ""+empresa.id).findList();
+            System.out.println("");
+            System.out.println("id: "+empresa.id);
+            System.err.println("TAM "+formEmpresa.size());
+            return ok(perfilEmpresa.render("", formEmpresa, empresa));
+       }
+       return redirect(routes.ApplicationEmpresa.home());
     }
     //METODO PARA ELIMINAR FORMULARIO EMPRESAS
     public Result eliminarFormularioEmpresa(Long id) {
         RegistroEmpresa empresa = new EmpresaSession().getRegistroEmpresa();
-        List<FormularioEmpresa> instancia = FormularioEmpresa.find.where().ilike("registroEmpresa",""+empresa.id).findList();
-            for(FormularioEmpresa femp: instancia){
-                femp.delete();
-            }
-        return redirect(routes.ControllerEmpresa.listarFormularioEmpresa());
+        if(empresa!=null){
+            List<FormularioEmpresa> instancia = FormularioEmpresa.find.where().ilike("registroEmpresa",""+empresa.id).findList();
+                for(FormularioEmpresa femp: instancia){
+                    femp.delete();
+                }
+            return redirect(routes.ControllerEmpresa.listarFormularioEmpresa());
+        }
+        return redirect(routes.ApplicationEmpresa.home());
     } 
     //METODO GET PARA EDITAR EL FORMULARIO DE UNA EMPRESA
     public Result editarPerfilEmpresaGet(Long id){
@@ -60,25 +66,29 @@ public class ControllerEmpresa extends Controller{
          return ok(formularioEmpresa.render("Formulario Empresa", formEmp,
                   routes.ControllerEmpresa.editarPerfilEmpresaPost(id)));
      }
-    //METODO POST PARA EDITAR EL FORMULARIO DE UN ESTUDIANTE
+    //METODO POST PARA EDITAR EL FORMULARIO DE LA EMPRESA
      public Result editarPerfilEmpresaPost(Long id){
          FormularioEmpresa instancia = FormularioEmpresa.find.byId(id);
          Form<FormularioEmpresa> formEmp = formFactory.form(FormularioEmpresa.class).fill(instancia).bindFromRequest();
+          RegistroEmpresa empresa = new EmpresaSession().getRegistroEmpresa();
+        if(empresa!=null){
          if(formEmp.hasErrors()){
-             return badRequest(formularioEmpresa.render("Encontramos errores en el formulario", formEmp, 
-                    routes.ControllerEmpresa.editarPerfilEmpresaPost(id)));
-         }
-         FormularioEmpresa formEmpresa = formEmp.get();
-         instancia.nombre = formEmpresa.nombre;
-         instancia.correoEmpresa = formEmpresa.correoEmpresa;
-         instancia.direccion = formEmpresa.direccion;
-         instancia.estadoContrataciones = formEmpresa.estadoContrataciones;
-         instancia.otrasContrataciones = formEmpresa.otrasContrataciones;
-         instancia.perfilEmpresarial = formEmpresa.perfilEmpresarial;
-         instancia.telefonoContacto = formEmpresa.telefonoContacto;
-         
-         instancia.save();
-         return redirect(routes.ControllerEmpresa.listarFormularioEmpresa());
+                return badRequest(formularioEmpresa.render("Encontramos errores en el formulario", formEmp, 
+                       routes.ControllerEmpresa.editarPerfilEmpresaPost(id)));
+            }
+            FormularioEmpresa formEmpresa = formEmp.get();
+            instancia.nombre = formEmpresa.nombre;
+            instancia.correoEmpresa = formEmpresa.correoEmpresa;
+            instancia.direccion = formEmpresa.direccion;
+            instancia.estadoContrataciones = formEmpresa.estadoContrataciones;
+            instancia.otrasContrataciones = formEmpresa.otrasContrataciones;
+            instancia.perfilEmpresarial = formEmpresa.perfilEmpresarial;
+            instancia.telefonoContacto = formEmpresa.telefonoContacto;
+
+            instancia.save();
+            return redirect(routes.ControllerEmpresa.listarFormularioEmpresa());
+        }
+         return redirect(routes.ApplicationEmpresa.home());
      }
      
      //METODO GET REGISTRO EMPRESA
